CMakeFilestxt odified for precompiled headers in libs, common sexpr and eeschema
Two header files  : the first has : Glib-2.0 (stl, etc...), gdk, gtk-3.0,
python-3.8 bindings.
The other one is identical, plus wxWidgets-3.0.
The local kikad GLEW includes are also manually inserted , taking care of the
insertion order.
Except for the local GLEW lib, no local header is included in the precompiled
header

A python tool is used to extract the included files with the help of gcc -H.
Currently, this is a HACK that implies a lot of manual tweaking.
A Cmake target/dependancy is planed to generate the two header files.